Break up isotopes mod to smaller files. Somehow breaks dep warnings
I tried fiddling with stacklevel parameter but no impact
